Split contact editor to separate forms for better manageable codes.
